requirements.txt
aiofiles==24.1.0
certifi==2025.7.14
fastapi==0.116.1
ffmpeg_python==0.2.0
httpx==0.28.1
numpy==1.24.3
openai==1.97.1
opencv_python==4.8.1.78
Pillow==11.3.0
pydantic==2.11.7
pydub==0.25.1
pycaps
python-dotenv==1.1.1
redis==5.0.1
stripe==12.3.0
supabase==2.17.0
uvicorn==0.35.0
yt_dlp==2025.7.21
